fixed sag fault; added mutex into hashmap model store; fixed np.float32 bug in msg helper

+++ metisfl/learner/message_helper.py
@@ -44,7 +44,7 @@ class MessageHelper:
             tensor = model.tensors.add()
             tensor.length = weight.size
             tensor.dimensions.extend(weight.shape)
-            weight = weight.astype(np.float32)
+            weight = weight.astype(np.float64)
 
             if self.scheme is not None:
                 model.encrypted = True
@@ -81,14 +81,14 @@ class MessageHelper:
                 weights.append(
                     np.array(
                         decrypted,
-                        dtype=np.float32
+                        dtype=np.float64
                     ).reshape(tensor.dimensions)
                 )
             else:
                 weights.append(
                     np.frombuffer(
                         buffer=tensor.value,
-                        dtype=np.float32,
+                        dtype=np.float64,
                         count=tensor.length
                     ).reshape(tensor.dimensions)
                 )
